# docs/openAPI_document/streamflix_openapi.yaml
openapi: 3.1.0
info:
  title: StreamFlix API
  version: v1
servers:
  - url: http://localhost:8080
    description: Generated server url
security:
  - bearerAuth: []
...
paths:
...
  /viewing-behaviour/watch:
    post:
      tags:
        - viewing-behaviour
      summary: Record watch event
      description: Record a viewing event with progress. Supports JSON, XML, CSV.
      operationId: watch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWatchEventRequest'
        required: true
      responses:
        '201':
          description: Watch event recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /titles:
    get:
      tags:
        - titles
      summary: Get all titles
      description: Get all movies and series. Supports JSON, XML, CSV.
      operationId: getTitles
      responses:
        '200':
          description: Titles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - titles
      summary: Create title
      description: Create a new movie or series
      operationId: createTitle
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTitleRequest'
        required: true
      responses:
        '201':
          description: Title created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TitleResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/TitleResponse'
        '400':
          description: Invalid input - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /viewing-behaviour/progress/{titleId}:
    get:
      tags:
        - viewing-behaviour
      summary: Get viewing progress
      description: Get viewing progress for a specific title. Supports JSON, XML, CSV.
      operationId: getViewingProgress
      parameters:
        - name: titleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Viewing progress retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/WatchEventResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No viewing progress found for this title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /viewing-behaviour/history:
    get:
      tags:
        - viewing-behaviour
      summary: Get watch history
      description: Get profile's viewing history. Supports JSON, XML, CSV.
      operationId: getWatchHistory
      parameters:
        - name: profileId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Watch history retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchEventResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchEventResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatchEventResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /titles/{id}:
    get:
      tags:
        - titles
      summary: Get title by ID
      description: Get specific movie or series. Supports JSON, XML, CSV.
      operationId: getTitleById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Title retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TitleResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TitleResponse'
            text/csv:
              schema:
                $ref: '#/components/schemas/TitleResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Title not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /titles/tvmaze/search:
    get:
      tags:
        - titles
      summary: Search shows via TVmaze
      description: >-
        Search for TV shows using the external TVmaze public API (https://api.tvmaze.com). Returns show details such as
        name, status, rating, and premiere date.
      operationId: searchTvMaze
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results returned (may be empty if no match)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TvMazeShowResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TvMazeShowResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /titles/tvmaze/lookup:
    get:
      tags:
        - titles
      summary: Look up a single show via TVmaze
      description: Fetch detailed information for a single show by name from the external TVmaze public API.
      operationId: lookupTvMaze
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Show details returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TvMazeShowResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/TvMazeShowResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: No show found on TVmaze for the given query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /titles/for-profile/{profileId}:
    get:
      tags:
        - titles
      summary: Get titles for profile
      description: Get titles filtered by profile age restrictions and viewing preferences.
      operationId: getTitlesForProfile
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Filtered titles retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
            text/csv:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TitleResponse'
        '401':
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
            application/xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    ErrorResponse:
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: string
...
    CreateWatchEventRequest:
      properties:
        titleId:
          type: string
          maxLength: 64
          minLength: 0
        episodeId:
          type: string
          maxLength: 64
          minLength: 0
        autoContinued:
          type: boolean
        profileId:
          type: string
        progressSeconds:
          type: integer
          format: int32
          minimum: 0
        finished:
          type: boolean
      required:
        - finished
        - profileId
        - titleId
    WatchEventResponse:
      properties:
        titleId:
          type: string
        titleName:
          type: string
        episodeId:
          type: string
        autoContinued:
          type: boolean
        startedAt:
          type: integer
          format: int64
        progressSeconds:
          type: integer
          format: int32
        finished:
          type: boolean
        lastUpdated:
          type: integer
          format: int64
    CreateTitleRequest:
      properties:
        name:
          type: string
          maxLength: 200
          minLength: 0
        description:
          type: string
          maxLength: 2000
          minLength: 0
        releaseYear:
          type: integer
          format: int32
          minimum: 1888
        durationSeconds:
          type: integer
          format: int32
          minimum: 1
        type:
          type: string
          enum:
            - MOVIE
            - SERIES
        genre:
          type: string
          enum:
            - ACTION
            - COMEDY
            - DRAMA
            - HORROR
            - THRILLER
            - ROMANCE
            - DOCUMENTARY
            - ANIMATION
            - SCIENCE_FICTION
            - FANTASY
            - CRIME
            - MYSTERY
        maturityRating:
          type: string
          enum:
            - ALL
            - PG
            - TEEN
            - MATURE
        contentWarnings:
          type: array
          items:
            type: string
            enum:
              - VIOLENCE
              - FEAR
              - COARSE_LANGUAGE
              - NUDITY
              - DRUG_USE
              - SEXUAL_CONTENT
          uniqueItems: true
        supportedQualities:
          type: array
          items:
            type: string
            enum:
              - SD
              - HD
              - UHD
          uniqueItems: true
      required:
        - name
        - releaseYear
        - supportedQualities
        - type
    TitleResponse:
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        releaseYear:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        type:
          type: string
          enum:
            - MOVIE
            - SERIES
        genre:
          type: string
          enum:
            - ACTION
            - COMEDY
            - DRAMA
            - HORROR
            - THRILLER
            - ROMANCE
            - DOCUMENTARY
            - ANIMATION
            - SCIENCE_FICTION
            - FANTASY
            - CRIME
            - MYSTERY
        maturityRating:
          type: string
          enum:
            - ALL
            - PG
            - TEEN
            - MATURE
        contentWarnings:
          type: array
          items:
            type: string
            enum:
              - VIOLENCE
              - FEAR
              - COARSE_LANGUAGE
              - NUDITY
              - DRUG_USE
              - SEXUAL_CONTENT
          uniqueItems: true
        supportedQualities:
          type: array
          items:
            type: string
            enum:
              - SD
              - HD
              - UHD
          uniqueItems: true
...
    ImageInfo:
      properties:
        medium:
          type: string
        original:
          type: string
    NetworkInfo:
      properties:
        name:
          type: string
    RatingInfo:
      properties:
        average:
          type: number
          format: double
    TvMazeShowResponse:
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        status:
          type: string
        type:
          type: string
        language:
          type: string
        rating:
          $ref: '#/components/schemas/RatingInfo'
        network:
          $ref: '#/components/schemas/NetworkInfo'
        image:
          $ref: '#/components/schemas/ImageInfo'
        summary:
          type: string
        premiered:
          type: string
        officialSite:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT